Cleaning up delay handler, and support animoto embeds
