+++ scripts/aem.js
@@ -750,32 +750,6 @@ async function waitForLCP(lcpBlocks) {
   });
 }
 
-const formConfigurations = [];
-
-/**
- * Checks if there are any form configurations.
- * @returns {boolean} True if there are form configurations, otherwise false.
- */
-function hasForms() {
-  return formConfigurations.length > 0;
-}
-
-function addFormConfiguration(formConfig) {
-  formConfigurations.push(formConfig);
-}
-
-function buildForms(hbspt) {
-  formConfigurations.forEach((formConfiguration) => {
-    const config = {
-      region: 'na1',
-      portalId: '21555329',
-      formId: formConfiguration.formId,
-      target: `.hubspot-form#${formConfiguration.targetElementId}`,
-    };
-    hbspt.forms.create(config);
-  });
-}
-
 init();
 
 export {
@@ -804,7 +778,4 @@ export {
   updateSectionsStatus,
   waitForLCP,
   wrapTextNodes,
-  hasForms,
-  addFormConfiguration,
-  buildForms,
 };
